cleanup and simplify extract fields
